refactor: FavoriteControl 컴포넌트 분리
